# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MLK Pro API",
    "version": "v1",
    "description": "Auto-generated from routes/api.php (internal API v1)."
  },
  "servers": [
    {
      "url": "http://localhost",
      "description": "Local dev"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/v1/team": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "TeamMemberController@index",
        "operationId": "TeamMemberController_index_get",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TeamMemberController@index",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:team_members"
        ]
      },
      "post": {
        "tags": [
          "team"
        ],
        "summary": "TeamMemberController@store",
        "operationId": "TeamMemberController_store_post",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TeamMemberController@store",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:team_members"
        ]
      }
    },
    "/api/v1/team/{teamMember}": {
      "put": {
        "tags": [
          "team"
        ],
        "summary": "TeamMemberController@update",
        "operationId": "TeamMemberController_update_put",
        "parameters": [
          {
            "name": "teamMember",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TeamMemberController@update",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:team_members"
        ]
      },
      "delete": {
        "tags": [
          "team"
        ],
        "summary": "TeamMemberController@destroy",
        "operationId": "TeamMemberController_destroy_delete",
        "parameters": [
          {
            "name": "teamMember",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-action": "App\\Http\\Controllers\\TeamMemberController@destroy",
        "x-middleware": [
          "api",
          "Illuminate\\Auth\\Middleware\\Authenticate:sanctum",
          "App\\Http\\Middleware\\EnsureInternalUser",
          "App\\Http\\Middleware\\EnsureNotSuspended",
          "App\\Http\\Middleware\\EnsureOnboardingIsComplete",
          "App\\Http\\Middleware\\EnsureCompanyFeature:team_members"
        ]
      }
    },
...
  }
}